+++ include/ambition/7kaaInterface/input.hh
@@ -41,6 +41,8 @@ enum class Action {
   Common_Confirm,
   Common_Reward,
 
+  Speed_Decrease_Standard,
+  Speed_Increase_Standard,
   Speed_SetToUnlimited,
   Speed_TogglePause,