+++ src/OMOUSE.cpp
@@ -234,6 +234,8 @@ void Mouse::init_key()
 	bind_key(KEYEVENT_CANCEL, "Escape");
 	bind_key(KEYEVENT_CONFIRM, "Return");
 
+	bind_key(KEYEVENT_SPEED_DECREASE_STANDARD, "-");
+	bind_key(KEYEVENT_SPEED_INCREASE_STANDARD, "=");
 	bind_key(KEYEVENT_SPEED_SET_TO_UNLIMITED, "\\");
 
 	bind_key(KEYEVENT_UNIT_REWARD, "W");