+++ src/ambition/7kaaInterface/input.cc
@@ -185,6 +185,8 @@ const std::map<Action, KeyEventType> KEY_EVENT_MAP = {
   { Action::Common_Confirm, KEYEVENT_CONFIRM },
   { Action::Common_Reward, KEYEVENT_UNIT_REWARD },
 
+  { Action::Speed_Decrease_Standard, KEYEVENT_SPEED_DECREASE_STANDARD },
+  { Action::Speed_Increase_Standard, KEYEVENT_SPEED_INCREASE_STANDARD },
   { Action::Speed_SetToUnlimited, KEYEVENT_SPEED_SET_TO_UNLIMITED },
   { Action::Speed_TogglePause, KEYEVENT_SPEED_TOGGLE_PAUSE },
 
@@ -798,6 +800,24 @@ bool detectSpeedChangeKeys(
 
   constexpr auto STEP_STANDARD = 3;
   constexpr auto SPEED_UNLIMITED = 99;
+
+  if (keyCode == getKeyEvent(Action::Speed_Decrease_Standard)) {
+    /* sys.set_speed(0) actually toggles between paused and not, so we do not
+     * want to call it if we are already paused. */
+    if (config.frame_speed > 0) {
+      const auto step = STEP_STANDARD;
+      sys.set_speed(std::max(0, config.frame_speed - step));
+      sys.user_pause_flag = !config.frame_speed;
+    }
+    return true;
+  }
+
+  if (keyCode == getKeyEvent(Action::Speed_Increase_Standard)) {
+    const auto step = STEP_STANDARD;
+    sys.set_speed(std::min(SPEED_UNLIMITED, config.frame_speed + step));
+    sys.user_pause_flag = 0;
+    return true;
+  }
 
   if (keyCode == getKeyEvent(Action::Speed_SetToUnlimited)) {
     sys.set_speed(SPEED_UNLIMITED);